--- Второй семестр/README.md
@@ -1,30 +1,0 @@
-# Лабораторная работа №1
-
-## Цель работы
-Исследовать влияние различных интервальных последовательностей на скорость выполнения алгоритма сортировки Шелла.
-
-## Методы
-Были реализованы три интервальные последовательности:
-1. **Последовательность по умолчанию**: `{5, 3, 1}`.
-2. **Последовательность Кнута**: `h = 3*h + 1`.
-3. **Последовательность Седжвика**: `h = 9 * (4^k - 2^k) + 1` или `4^k + 3 * 2^(k-1) + 1`.
-
-## Результаты
-
-### Таблица с ключевыми показателями
-
-| Название способа       | Формула интервалов | Размер массива | Время сортировки (нс) | Вывод                                                                 |
-|------------------------|--------------------|----------------|-----------------------|-----------------------------------------------------------------------|
-| Сортировка Шелла (Default) | {5, 3, 1}          | 20             | 12000                 | Простая последовательность, подходит для малых массивов.              |
-| Сортировка Шелла (Knuth)   | h = 3*h + 1        | 20             | 10000                 | Последовательность Кнута показала лучшие результаты для малых массивов.|
-| Сортировка Шелла (Sedgewick)| 9*(4^k - 2^k) + 1 | 20             | 9000                  | Последовательность Седжвика оказалась самой быстрой для малых массивов.|
-
-## Выводы
-1. **Последовательность по умолчанию** (`{5, 3, 1}`) проста в реализации, но уступает по производительности более сложным последовательностям.
-2. **Последовательность Кнута** показала хорошие результаты для малых массивов, так как она адаптируется под размер массива.
-3. **Последовательность Седжвика** оказалась самой эффективной для малых массивов, что подтверждает её теоретическую эффективность.
-
-## Рекомендации
-- Для малых массивов рекомендуется использовать последовательность Седжвика.
-- Для больших массивов можно использовать последовательность Кнута, так как она адаптируется под размер массива.
-- Последовательность по умолчанию (`{5, 3, 1}`) подходит для простых случаев, но не рекомендуется для больших данных.
